// autogpt_platform/frontend/src/app/api/__generated__/models/credentialsMetaResponseType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */

export type CredentialsMetaResponseType =
  (typeof CredentialsMetaResponseType)[keyof typeof CredentialsMetaResponseType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CredentialsMetaResponseType = {
  api_key: "api_key",
  oauth2: "oauth2",
  user_password: "user_password",
  host_scoped: "host_scoped",
} as const;